pyproject.toml
[project]
name = "gw-team"
version = "0.1.0"
description = ""
authors = [
    {name = "franklin",email = "frank.fkn.23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.5"
taskipy = "^1.14.1"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
factory-boy = "^3.3.3"
freezegun = "^1.5.3"
testcontainers = "^4.12.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']
fix = true

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function' # aparentemente temporario

[tool.taskipy.tasks]
pre_fmt = 'ruff check'
fmt = 'ruff format'
run = 'fastapi dev gw_team/app.py'
pre_tst = 'task fmt'
tst = 'pytest -s -x --cov=gw_team --cov-config=.coveragerc -vv'
post_tst = 'coverage html'
